Pull out common functions from integration test
