+++ contracts/hackatom/Cargo.toml
@@ -23,8 +23,8 @@ overflow-checks = true
 
 [dependencies]
 cosmwasm = { path = "../.." }
-serde_json = { version = "1.0.39" }
-serde = { version = "1.0", default-features = false, features = ["derive"] }
+serde_json = { version = "1.0.41" }
+serde = { version = "1.0.60", default-features = false, features = ["derive"] }
 failure = "0.1.5"
 
 [dev-dependencies]
